Move ORM models into the blueprints that use them.
Farewell nitre we hardly knew ye.
