+++ vulpes/blueprints/snapcast/sql.py
@@ -1,13 +1,13 @@
 from datetime import datetime, timezone
 from uuid import UUID
 
-from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import update
 
-from ...nitre import Podcast
+from ... import db
+from .models import Podcast
 
 
-def touch_podcast(db: SQLAlchemy, podcast_uuid: UUID):
+def touch_podcast(podcast_uuid: UUID):
     """Update the last_modified field for a podcast. Called on cache-invalidating requests."""
     db.session.execute(
         update(Podcast)
