+++ vulpes/blueprints/snapcast/views.py
@@ -2,24 +2,21 @@ from datetime import datetime, timedelta, timezone
 from uuid import UUID, uuid4
 
 from flask import Response, jsonify, request
-from flask_sqlalchemy import SQLAlchemy
 from podgen import Category, Episode, Media, Person, Podcast
 from sqlalchemy import select
 
-from ... import nitre
-from ...connections import uses_db
-from . import bp
+from ... import db
+from . import bp, models
 from .decorators import authorization_required
 from .sql import touch_podcast
 
 
 @bp.route("/<uuid:podcast_uuid>/feed.xml", methods=["GET"])
-@uses_db
-def generate_feed(db: SQLAlchemy, podcast_uuid: UUID):
+def generate_feed(podcast_uuid: UUID):
     """Pull podcast and episode data from the db and generate a podcast xml file."""
-    cast: nitre.Podcast = db.first_or_404(
-        select(nitre.Podcast)
-        .where(nitre.Podcast.uuid == podcast_uuid),
+    cast: models.Podcast = db.first_or_404(
+        select(models.Podcast)
+        .where(models.Podcast.uuid == podcast_uuid),
     )
 
     # The caveat to sqlalchemy and sqlite: It stores datetimes as naive.
@@ -44,8 +41,8 @@ def generate_feed(db: SQLAlchemy, podcast_uuid: UUID):
     )
 
     episodes = db.session.scalars(
-        select(nitre.Episode)
-        .where(nitre.Episode.podcast_uuid == podcast_uuid),
+        select(models.Episode)
+        .where(models.Episode.podcast_uuid == podcast_uuid),
     )
     for episode in episodes:
         e = Episode(
@@ -72,15 +69,14 @@ def generate_feed(db: SQLAlchemy, podcast_uuid: UUID):
 
 
 @bp.route("/<uuid:podcast_uuid>/feed.xml", methods=["HEAD"])
-@uses_db
-def feed_head(db: SQLAlchemy, podcast_uuid: UUID):
+def feed_head(podcast_uuid: UUID):
     """Set headers for a HEAD request to a feed.
 
     Fill `Last-Modified` to save on data transfer.
     """
     last_modified = db.one_or_404(
-        select(nitre.Podcast.last_modified)
-        .where(nitre.Podcast.uuid == podcast_uuid),
+        select(models.Podcast.last_modified)
+        .where(models.Podcast.uuid == podcast_uuid),
     )
     response = Response()
     response.last_modified = last_modified.replace(tzinfo=timezone.utc)
@@ -96,9 +92,8 @@ def generate_snapcast():
 
 
 @bp.route("/<uuid:podcast_uuid>/publish", methods=["POST"])
-@uses_db
 @authorization_required
-def publish_episode(db: SQLAlchemy, podcast_uuid: UUID):
+def publish_episode(podcast_uuid: UUID):
     """Add a new episode to a podcast.
 
     Required elements in JSON request body:
@@ -134,7 +129,7 @@ def publish_episode(db: SQLAlchemy, podcast_uuid: UUID):
         "pub_date":         pub_date,
     }
 
-    db.session.add(nitre.Episode(**data))
-    touch_podcast(db, podcast_uuid)
+    db.session.add(models.Episode(**data))
+    touch_podcast(podcast_uuid)
     db.session.commit()
     return jsonify(success=True)
